Refactor: money.rs out of lib.rs
